[python/asyncio] explicitly close client session via async context manager
