+++ modules/openapi-generator/src/main/resources/python/api_client.mustache
@@ -76,13 +76,25 @@ class ApiClient(object):
         self.user_agent = '{{#httpUserAgent}}{{{.}}}{{/httpUserAgent}}{{^httpUserAgent}}OpenAPI-Generator/{{{packageVersion}}}/python{{/httpUserAgent}}'
         self.client_side_validation = configuration.client_side_validation
 
+    {{#asyncio}}
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, exc_type, exc_value, traceback):
+        await self.close()
+    {{/asyncio}}
+    {{^asyncio}}
     def __enter__(self):
         return self
 
     def __exit__(self, exc_type, exc_value, traceback):
         self.close()
+    {{/asyncio}}
 
-    def close(self):
+    {{#asyncio}}async {{/asyncio}}def close(self):
+        {{#asyncio}}
+        await self.rest_client.close()
+        {{/asyncio}}
         if self._pool:
             self._pool.close()
             self._pool.join()
