+++ modules/openapi-generator/src/main/resources/python/asyncio/rest.mustache
@@ -10,7 +10,6 @@ import ssl
 
 import aiohttp
 import certifi
-import asyncio
 # python 2 and python 3 compatibility library
 from six.moves.urllib.parse import urlencode
 
@@ -77,8 +76,8 @@ class RESTClientObject(object):
                 connector=connector
             )
 
-    def __del__(self):
-        asyncio.ensure_future(self.pool_manager.close())
+    async def close(self):
+        await self.pool_manager.close()
 
     async def request(self, method, url, query_params=None, headers=None,
                       body=None, post_params=None, _preload_content=True,
