+++ samples/client/petstore/python-asyncio/petstore_api/rest.py
@@ -18,7 +18,6 @@ import ssl
 
 import aiohttp
 import certifi
-import asyncio
 # python 2 and python 3 compatibility library
 from six.moves.urllib.parse import urlencode
 
@@ -85,8 +84,8 @@ class RESTClientObject(object):
                 connector=connector
             )
 
-    def __del__(self):
-        asyncio.ensure_future(self.pool_manager.close())
+    async def close(self):
+        await self.pool_manager.close()
 
     async def request(self, method, url, query_params=None, headers=None,
                       body=None, post_params=None, _preload_content=True,
