+++ samples/client/petstore/python-asyncio/tests/test_pet_api.py
@@ -18,22 +18,13 @@ import petstore_api
 from petstore_api import Configuration
 from petstore_api.rest import ApiException
 
-from .util import id_gen
+from .util import id_gen, async_test
 
 import json
 
 import urllib3
 
 HOST = 'http://localhost:80/v2'
-
-
-def async_test(f):
-    def wrapper(*args, **kwargs):
-        coro = asyncio.coroutine(f)
-        future = coro(*args, **kwargs)
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(future)
-    return wrapper
 
 
 class TestPetApiTests(unittest.TestCase):
