+++ samples/client/petstore/python-asyncio/tests/util.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 
+import asyncio
 import random
 
 
@@ -8,4 +9,10 @@ def id_gen(bits=32):
     return int(random.getrandbits(bits))
 
 
-
+def async_test(f):
+    def wrapper(*args, **kwargs):
+        coro = asyncio.coroutine(f)
+        future = coro(*args, **kwargs)
+        loop = asyncio.get_event_loop()
+        loop.run_until_complete(future)
+    return wrapper
